+++ mediation/Vungle/source/plugin/Assets/GoogleMobileAds/Common/Mediation/Vungle/IVungleClient.cs
@@ -20,10 +20,8 @@ namespace GoogleMobileAds.Common.Mediation.Vungle
 {
     public interface IVungleClient
     {
-        void UpdateConsentStatus(VungleConsent consentStatus);
-        void UpdateConsentStatus(VungleConsent consentStatus,
+        void UpdateConsentStatus(VungleConsentStatus consentStatus,
                                  String consentMessageVersion);
-        VungleConsent GetCurrentConsentStatus();
-        String GetCurrentConsentMessageVersion();
+        void UpdateCCPAStatus(VungleCCPAStatus consentStatus);
     }
 }